// src/api/generated/model/roomsControllerGetRoomPromoCodeUsagesParams.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

export type RoomsControllerGetRoomPromoCodeUsagesParams = {
/**
 * ID амбассадоров для фильтрации
 */
ambassadorId?: string[];
/**
 * ID событий для фильтрации
 */
eventId?: string[];
/**
 * ID спринтов для фильтрации
 */
sprintId?: string[];
/**
 * Дата начала периода (включительно)
 */
dateFrom?: string;
/**
 * Дата окончания периода (включительно)
 */
dateTo?: string;
/**
 * Номер страницы (от 1)
 */
page?: number;
/**
 * Размер страницы
 */
size?: number;
};